compensated_readings: make 'timestamp' tz-aware. (#15)
* compensated_readings: make 'timestamp' tz-aware.

The 'timestamp' object in the compensated_readings class was prevously
"naive", meaning it did not have any timezone information and was
therefore ambiguous.  This change explicitly sets the timezone to UTC.

* Restore Python <3.6 compatibility by using pytz's UTC.

* setup.py: remove source-level dep on bme280 module.

Since bme280 now includes dependencies (pytz) outside the stdlib,
it can't be directly imported in the setup.py script. Instead, we now
read the __version__ by reading the __init__.py file directly.

* Also clean up timezone imports on test_bme280.py.